Fix test for preview
